+++ packages/hashtag/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@draft-js-plugins/hashtag",
-  "version": "4.0.0-beta2",
+  "version": "4.0.0-beta3",
   "description": "Hashtag Plugin for DraftJS",
   "author": {
     "name": "Nik Graf",
